+++ libft/ft_putstr_fd.c
@@ -6,7 +6,7 @@
 /*   By: kasen <kasen@student.42istanbul.com.tr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/08/10 15:11:00 by kasen             #+#    #+#             */
-/*   Updated: 2026/08/20 20:20:11 by kasen            ###   ########.fr       */
+/*   Updated: 2026/08/31 14:35:48 by kasen            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putstr_fd(char *s, int fd)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
